HDDS-13160. Suspend should wait
