+++ hadoop-ozone/ozone-manager/src/main/java/org/apache/hadoop/ozone/om/service/SnapshotDeletingService.java
@@ -35,7 +35,6 @@ import java.util.Objects;
 import java.util.Optional;
 import java.util.UUID;
 import java.util.concurrent.TimeUnit;
-import java.util.concurrent.atomic.AtomicBoolean;
 import java.util.concurrent.atomic.AtomicLong;
 import java.util.stream.Collectors;
 import org.apache.hadoop.hdds.conf.OzoneConfiguration;
@@ -87,7 +86,6 @@ public class SnapshotDeletingService extends AbstractKeyDeletingService {
   private final OzoneManager ozoneManager;
   private final OmSnapshotManager omSnapshotManager;
   private final SnapshotChainManager chainManager;
-  private final AtomicBoolean suspended;
   private final OzoneConfiguration conf;
   private final AtomicLong successRunCount;
   private final int keyLimitPerTask;
@@ -107,7 +105,6 @@ public class SnapshotDeletingService extends AbstractKeyDeletingService {
         ozoneManager.getMetadataManager();
     this.chainManager = omMetadataManager.getSnapshotChainManager();
     this.successRunCount = new AtomicLong(0);
-    this.suspended = new AtomicBoolean(false);
     this.conf = ozoneManager.getConfiguration();
     this.snapshotDeletionPerTask = conf.getInt(SNAPSHOT_DELETING_LIMIT_PER_TASK,
         SNAPSHOT_DELETING_LIMIT_PER_TASK_DEFAULT);
@@ -319,26 +316,6 @@ public class SnapshotDeletingService extends AbstractKeyDeletingService {
     BackgroundTaskQueue queue = new BackgroundTaskQueue();
     queue.add(new SnapshotDeletingTask());
     return queue;
-  }
-
-  private boolean shouldRun() {
-    return !suspended.get() && ozoneManager.isLeaderReady();
-  }
-
-  /**
-   * Suspend the service.
-   */
-  @VisibleForTesting
-  public void suspend() {
-    suspended.set(true);
-  }
-
-  /**
-   * Resume the service if suspended.
-   */
-  @VisibleForTesting
-  public void resume() {
-    suspended.set(false);
   }
 
   public long getSuccessfulRunCount() {
